+++ src/common/models/target.js
@@ -1,8 +1,9 @@
 export class Target {
-    constructor(id, name, pattern, selector) {
+    constructor(id, name, pattern, selector, isActive) {
         this.id = id;
         this.name = name;
         this.pattern = pattern;
         this.selector = selector;
+        this.isActive = isActive;
     }
 }