+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2023-03-25
+
+### Added
+
+- Disable unnecessary targets
+- `0 sec` autoclose time option
+
 ## [0.1.0] - 2023-02-12
 
 ### Added
@@ -16,5 +23,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use regular expression for URL mathing
 - Configure popup autoclosing and autoclosing time
 
-[Unreleased]: https://github.com/vikian050194/click/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/vikian050194/click/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/vikian050194/click/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/vikian050194/click/releases/tag/v0.1.0